# Wallet/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SmartBank Wallet API Documentation",
    "version": "1.0.0",
    "description": "Dokumentasi API E-Wallet SmartBank (Tier-2) yang terintegrasi secara microservice dengan Central Bank moneter. Gunakan fitur 'Try it out' untuk menguji endpoint secara langsung.\n\n## Autentikasi\nSemua endpoint protected memerlukan header `Authorization: Bearer <accessToken>` yang didapat dari `POST /api/v1/auth/login`.\n\n## Keamanan Transaksi\n- **PIN Transaksi**: Endpoint finansial (transfer, bayar tagihan, topup, withdraw, claim stimulus) memerlukan header `X-Wallet-Pin` berisi 6-digit PIN.\n- **Idempotency-Key**: Endpoint mutasi saldo (transfer, bayar tagihan, apply loan, repay loan) memerlukan header `Idempotency-Key` unik untuk mencegah double-spend.\n\n## Response Format\nSemua response mengikuti format standar:\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"error\": null,\n  \"meta\": { \"request_id\": \"...\", \"timestamp\": \"...\" }\n}\n```"
  },
  "servers": [
    {
      "url": "http://localhost:6969",
      "description": "Server Wallet Lokal (Host)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Masukkan token JWT akses (accessToken) yang didapatkan dari endpoint login. Format: Bearer <token>"
      },
      "walletPin": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Wallet-Pin",
        "description": "6-digit PIN transaksi pengguna"
      },
      "idempotencyKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key",
        "description": "Kunci unik untuk mencegah double-spend pada transaksi finansial"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/wallets/me/balance": {
      "get": {
        "tags": ["Wallet"],
        "summary": "Cek Saldo Dompet",
        "description": "Membaca saldo riil (CBDC_IDR), hold, dan limit transaksi harian terkini yang terdaftar di Central Bank untuk akun pengguna yang login.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Detail saldo berhasil ditarik." },
          "401": { "description": "Tidak diotorisasi (Token JWT tidak valid atau kosong)." }
        }
      }
    },
    "/api/v1/wallets/me/transactions": {
      "get": {
        "tags": ["Wallet"],
        "summary": "Riwayat Mutasi Rekening",
        "description": "Menampilkan seluruh daftar transaksi moneter (Transfer, Kredit, Pembayaran) beserta detail debit/kredit dan pihak lawan (double-entry ledger).",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Riwayat mutasi berhasil diambil." },
          "401": { "description": "Tidak diotorisasi." }
        }
      }
    },
...
  }
}